ada: Fix usage of Table.Table in Fmap
Table.Table can be instantiated to use either 0-based or 1-based
indexing, which can cause some confusion and make 0-based instances get
used as 1-based ones.

This was the case for two tables in Fmap before this patch. That did not
cause any bugs but allocated an extra cell in the arrays that went
unused.

This patch also replaces Increment_Last-and-assignment combos with
equivalent calls to Append.

gcc/ada/ChangeLog:

	* fmap.adb (File_Mapping, Path_Mapping): Fix instantiations.
	(Add_To_File_Map): Use Table.Table.Append.