ada: Fix usage of Table.Table in Sem_Ch12
Before this patch, Sem_Ch12 jumped through questionable hoops in the way
it used its Generics_Renaming table that involved defensive calls to the
'Valid attribute. No known bug has been caused by this, but valgrind
reported incorrect memory operations because of it.

After analysis, the problem seems to be a mix 0-based and 1-based
indexing in the uses of Generic_Renamings and a convoluted interface for
the Set_Instance_Of procedure, leading to an unclear status for
Generic_Renamings.Table (0).

This patch fixes those problems and removes the accompanying defensive
code.

gcc/ada/ChangeLog:

	* sem_ch12.adb (Build_Local_Package)
	(Analyze_Package_Instantiation, Analyze_Subprogram_Instantiation):
	Fix Set_Last calls.
	(Set_Instance_Of): Use Table.Table.Append.
	(Save_And_Reset): Remove useless call. Remove defensive code.
	(Restore): Remove incorrect Set_Last call and adapt to
	Set_Instance_Of change.