vect: Disable niters-based skipping of uncounted vectorized loops
The iteration count profitability check is irrelevant for uncounted
loops given that, even at runtime, the number of iterations is unknown
at the start of loop execution.

Likewise, the test for skipping the vectorized version of the loop is
based on whether the number of iterations that will be run for the
loop and whether it is smaller than the vectorization factor.  As this
is undetermined, the check can never be run for uncounted loops.

Consequently, we skip these checks.

gcc/ChangeLog:

	* tree-vect-loop-manip.cc (vect_do_peeling): Disable vector
	loop skip checking.
	(vect_loop_versioning): skip profitability check for uncounted loops.