tree-optimization/122079 - PRE antic compute doesn't converge
The following re-instantiates the pruning of the ANTIC_IN value set
by the previous iterations one by reverting part of
r16-3945-gc30f58c3f7ec25.  The earlier fixes made sure the initial
value set is appropriately big to not cause this to lose optimizations.
But it seems to be still required for correctness given iteration
order means we combine ANTIC_IN sets from different generations which
can be prone to oscillations in CFG cycles.

	PR tree-optimization/122079
	* tree-ssa-pre.cc (compute_antic_aux): Re-instantiate
	ANTIC_IN value pruning by the old solution.

	* gcc.dg/torture/pr122079-2.c: New testcase.
	* gcc.dg/torture/pr122079-3.c: Likewise.