vect: guard niters manipulation with `LOOP_VINFO_NITERS_UNCOUNTED_P'
In `vect_do_peeling' and `vect_transform_loop', there are several bits
of logic reliant on niters that need to be handled differently in the
case of uncounted loops.

Firstly When we peel the loop, adding a prolog,  we subtract the
prolog peeling factor from the original number of iterations for the
main loop.

Then, upon vectorization of the main loop, we need to update the
iteration upper-bound to reflect the fact that each iteration now acts
on VF elements, such that less iterations will be needed.

Both of these updates become unnecessary when we don't have an IV
counting exit.  Therefore, it is sufficient to guard these
manipulations behind a check for whether the loop we're dealing with
is uncounted.

gcc/ChangeLog:

	* tree-vect-loop-manip.cc (vect_do_peeling): Disable niters
	update.
	* tree-vect-loop.cc (vect_transform_loop): Likewise.