tree-optimization/105490 - improve COND_EXPR bool pattern
We miss to add a mask conversion from the mask producer to the
appropriate mask for the condition operation.  The following moves
required helpers and adds the missing part of the pattern.  That's
required both for the case we have different mask element sizes
and for the case we have a different number of elements because
cond expression vectorization doesn't handle the mask having
different nunits than the data vector.

	PR tree-optimization/105490
	* tree-vect-patterns.cc (build_mask_conversion): Move earlier.
	(vect_convert_mask_for_vectype): Likewise.
	(vect_recog_bool_pattern): Remove redundant truth type
	construction.  Add missing possibly required mask conversion.

	* gcc.dg/vect/vect-cond-14.c: New testcase.