c++: Fix ICE with concept used as function [PR116477]
As suggested by Patrick Palka in the bug report, the diagnostic check
for concept_check_p(fn) was being done too late in finish_call_expr(),
which led to an early return inside if (processing_template_decl), which
meant that the error wouldn't be triggered when we are in a type
dependence early exit. This fix makes sure that this error is handled in
the semantic analysis phase, and avoids the failed assertion later in
tsubst_expr().

	PR c++/116477

gcc/cp/ChangeLog:

	* semantics.cc (finish_call_expr): Move concept_check_p diagnostic
	before processing_template_decl check to catch errors earlier.

gcc/testsuite/ChangeLog:

	* g++.dg/cpp2a/concepts-pr116477.C: New test.

Signed-off-by: Egas Ribeiro <egas.g.ribeiro@tecnico.ulisboa.pt>