tree-optimization/122212 - fix CLZ detection
The following corrects a mistake with the zero value handling which
was broken because the bits bias was applied first which works
for the special-case using a bit-and but not when using a conditional
move.  Apply this after the fact instead where it also more easily
folds with an existing bias we compensate.

	PR tree-optimization/122212
	* tree-ssa-forwprop.cc (simplify_count_zeroes): Apply
	bias for CLZ after dealing with the zero special value.

	* gcc.dg/torture/pr122212.c: New testcase.